# sstable: add a writeQueue to the sstable writer