--- a/sstable/writer.go
+++ b/sstable/writer.go
@@ -11,6 +11,7 @@ import (
 	"fmt"
 	"io"
 	"math"
+	"sync"
 
 	"github.com/cespare/xxhash/v2"
 	"github.com/cockroachdb/errors"
@@ -109,6 +110,8 @@ type Writer struct {
 	successor               Successor
 	tableFormat             TableFormat
 	cache                   *cache.Cache
+	restartInterval         int
+	checksumType            ChecksumType
 	// disableKeyOrderChecks disables the checks that keys are added to an
 	// sstable in order. It is intended for internal use only in the construction
 	// of invalid sstables for testing. See tool/make_test_sstables.go.
@@ -170,6 +173,21 @@ type Writer struct {
 	// blockBuf consists of the state which is owned by and used by the Writer client
 	// goroutine.
 	blockBuf blockBuf
+
+	queueState struct {
+		// writeQueue is used to write data blocks to disk. The writeQueue is primarily
+		// used to maintain the order in which data blocks must be written to disk. For
+		// this reason, every single data block write must be done through the writeQueue.
+		writeQueue *writeQueue
+	}
+}
+
+var writeTaskPool = sync.Pool{
+	New: func() interface{} {
+		t := &writeTask{}
+		t.compressionDone = make(chan bool, 1)
+		return t
+	},
 }
 
 type checksummer struct {
@@ -228,16 +246,25 @@ type dataBlockBuf struct {
 	compressed []byte
 }
 
+var dataBlockBufPool = sync.Pool{
+	New: func() interface{} {
+		return &dataBlockBuf{}
+	},
+}
+
+func newDataBlockBuf(restartInterval int, checksumType ChecksumType) *dataBlockBuf {
+	d := dataBlockBufPool.Get().(*dataBlockBuf)
+	d.dataBlock.restartInterval = restartInterval
+	d.checksummer.checksumType = checksumType
+	return d
+}
+
 func (d *dataBlockBuf) finish() {
 	d.uncompressed = d.dataBlock.finish()
 }
 
 func (d *dataBlockBuf) compressAndChecksum(c Compression) {
 	d.compressed = compressAndChecksum(d.uncompressed, c, &d.blockBuf)
-}
-
-func (w *Writer) Error() error {
-	return w.err
 }
 
 type indexBlockAndBlockProperties struct {
@@ -255,8 +282,8 @@ type indexBlockAndBlockProperties struct {
 //
 // TODO(peter): untested
 func (w *Writer) Set(key, value []byte) error {
-	if err := w.Error(); err != nil {
-		return err
+	if w.err != nil {
+		return w.err
 	}
 	return w.addPoint(base.MakeInternalKey(key, 0, InternalKeyKindSet), value)
 }
@@ -267,8 +294,8 @@ func (w *Writer) Set(key, value []byte) error {
 //
 // TODO(peter): untested
 func (w *Writer) Delete(key []byte) error {
-	if err := w.Error(); err != nil {
-		return err
+	if w.err != nil {
+		return w.err
 	}
 	return w.addPoint(base.MakeInternalKey(key, 0, InternalKeyKindDelete), nil)
 }
@@ -280,8 +307,8 @@ func (w *Writer) Delete(key []byte) error {
 //
 // TODO(peter): untested
 func (w *Writer) DeleteRange(start, end []byte) error {
-	if err := w.Error(); err != nil {
-		return err
+	if w.err != nil {
+		return w.err
 	}
 	return w.addTombstone(base.MakeInternalKey(start, 0, InternalKeyKindRangeDelete), end)
 }
@@ -293,8 +320,8 @@ func (w *Writer) DeleteRange(start, end []byte) error {
 //
 // TODO(peter): untested
 func (w *Writer) Merge(key, value []byte) error {
-	if err := w.Error(); err != nil {
-		return err
+	if w.err != nil {
+		return w.err
 	}
 	return w.addPoint(base.MakeInternalKey(key, 0, InternalKeyKindMerge), value)
 }
@@ -306,8 +333,8 @@ func (w *Writer) Merge(key, value []byte) error {
 // point entries. Additionally, range deletion tombstones must be fragmented
 // (i.e. by keyspan.Fragmenter).
 func (w *Writer) Add(key InternalKey, value []byte) error {
-	if err := w.Error(); err != nil {
-		return err
+	if w.err != nil {
+		return w.err
 	}
 
 	switch key.Kind() {
@@ -511,8 +538,8 @@ func (w *Writer) RangeKeyDelete(start, end []byte) error {
 // overlap. Range keys may be added out of order relative to point keys and
 // range deletions.
 func (w *Writer) AddRangeKey(key InternalKey, value []byte) error {
-	if err := w.Error(); err != nil {
-		return err
+	if w.err != nil {
+		return w.err
 	}
 	return w.addRangeKey(key, value)
 }
@@ -741,32 +768,31 @@ func (w *Writer) maybeAddToFilter(key []byte) {
 	}
 }
 
+func (w *Writer) flush(key InternalKey) error {
+	w.dataBlockBuf.finish()
+	w.dataBlockBuf.compressAndChecksum(w.compression)
+
+	var err error
+	writeTask := writeTaskPool.Get().(*writeTask)
+	// We're setting compressionDone to indicate that compression of this block
+	// has already been completed.
+	writeTask.compressionDone <- true
+	writeTask.buf = w.dataBlockBuf
+	writeTask.indexSepKey = key
+
+	w.dataBlockBuf = nil
+	err = w.queueState.writeQueue.addSync(writeTask)
+	w.dataBlockBuf = newDataBlockBuf(w.restartInterval, w.checksumType)
+
+	return err
+}
+
 func (w *Writer) maybeFlush(key InternalKey, value []byte) error {
 	if !shouldFlush(key, value, &w.dataBlockBuf.dataBlock, w.blockSize, w.blockSizeThreshold) {
 		return nil
 	}
 
-	err := func() error {
-		w.dataBlockBuf.finish()
-		w.dataBlockBuf.compressAndChecksum(w.compression)
-
-		var bh BlockHandle
-		var err error
-		if bh, err = w.writeCompressedBlock(w.dataBlockBuf.compressed, w.dataBlockBuf.tmp[:]); err != nil {
-			return err
-		}
-
-		var bhp BlockHandleWithProperties
-		if bhp, err = w.maybeAddBlockPropertiesToBlockHandle(bh); err != nil {
-			return err
-		}
-
-		prevKey := base.DecodeInternalKey(w.dataBlockBuf.dataBlock.curKey)
-		if err = w.addIndexEntry(prevKey, key, bhp, w.dataBlockBuf.tmp[:]); err != nil {
-			return err
-		}
-		return nil
-	}()
+	err := w.flush(key)
 
 	if err != nil {
 		w.err = err
@@ -1015,7 +1041,11 @@ func (w *Writer) Close() (err error) {
 		w.syncer = nil
 	}()
 
-	if err := w.Error(); err != nil {
+	// finish must be called before we check for an error, because finish will
+	// block until every single task added to the writeQueue has been processed,
+	// and an error could be encountered while any of those tasks are processed.
+	w.err = w.queueState.writeQueue.finish()
+	if w.err != nil {
 		return err
 	}
 
@@ -1324,6 +1354,8 @@ func NewWriter(f writeCloseSyncer, o WriterOptions, extraOpts ...WriterOption) *
 		successor:               o.Comparer.Successor,
 		tableFormat:             o.TableFormat,
 		cache:                   o.Cache,
+		restartInterval:         o.BlockRestartInterval,
+		checksumType:            o.Checksum,
 		indexBlock: blockWriter{
 			restartInterval: 1,
 		},
@@ -1343,15 +1375,15 @@ func NewWriter(f writeCloseSyncer, o WriterOptions, extraOpts ...WriterOption) *
 		coalescer: rangekey.Coalescer{},
 	}
 
-	w.dataBlockBuf = &dataBlockBuf{
-		dataBlock: blockWriter{
-			restartInterval: o.BlockRestartInterval,
-		},
-	}
-	w.dataBlockBuf.checksummer = checksummer{checksumType: o.Checksum}
+	w.dataBlockBuf = newDataBlockBuf(w.restartInterval, w.checksumType)
+
 	w.blockBuf = blockBuf{
 		checksummer: checksummer{checksumType: o.Checksum},
 	}
+	// We're creating the writeQueue with a size of 0, because we won't be using
+	// the async queue until parallel compression is enabled, and we do all block
+	// writes through writeQueue.addSync.
+	w.queueState.writeQueue = newWriteQueue(0, w)
 
 	if f == nil {
 		w.err = errors.New("pebble: nil file")