--- a/sstable/writer_test.go
+++ b/sstable/writer_test.go
@@ -304,7 +304,7 @@ func (c *testBlockPropCollector) FinishTable(_ []byte) ([]byte, error) {
 	return nil, nil
 }
 
-func TestWriter_BlockProperties_Errors(t *testing.T) {
+func TestWriterBlockPropertiesErrors(t *testing.T) {
 	blockPropErr := errors.Newf("block property collector failed")
 	testCases := []blockPropErrSite{
 		errSiteAdd,
@@ -349,8 +349,8 @@ func TestWriter_BlockProperties_Errors(t *testing.T) {
 				require.NoError(t, err)
 				// Addition of a second key completes the first block.
 				w.Add(k2, v2)
-				require.Error(t, w.Error())
-				require.Equal(t, blockPropErr, w.Error())
+				require.Error(t, w.err)
+				require.Equal(t, blockPropErr, w.err)
 			case errSiteFinishIndex:
 				require.NoError(t, err)
 				// Addition of a second key completes the first block.
@@ -359,14 +359,14 @@ func TestWriter_BlockProperties_Errors(t *testing.T) {
 				// The index entry for the first block is added after the completion of
 				// the second block, which is triggered by adding a third key.
 				w.Add(k3, v3)
-				require.Error(t, w.Error())
-				require.Equal(t, blockPropErr, w.Error())
+				require.Error(t, w.err)
+				require.Equal(t, blockPropErr, w.err)
 			}
 
 			err = w.Close()
 			if tc == errSiteFinishTable {
 				require.Error(t, err)
-				require.Equal(t, blockPropErr, w.Error())
+				require.Equal(t, blockPropErr, w.err)
 			}
 		})
 	}