This pr adds a writeQueue to the sstable writer. The queue is used to
write data blocks to disk. Currently, the queue will run synchronously
with the Writer client goroutine.

A writeTask, along with a dataBlockBuf is passed to the writerQueue. The
Writer Client goroutine will block until the writeQueue finishes writing
the block to disk.