@bananabrick 2022-01-22
Posting some benchmark results:

benchstat comparison for all ycsb benchmarks run for 30 minutes, on the same machines as nightly benchmarks.

```
name                old ops/sec  new ops/sec  delta
ycsb/E/values=1024   69.6k ± 7%   69.0k ± 9%    ~     (p=0.841 n=5+5)
ycsb/A/values=1024   93.0k ± 1%   92.9k ± 1%    ~     (p=1.000 n=5+5)
ycsb/D/values=1024    112k ±17%    109k ±16%    ~     (p=0.690 n=5+5)
ycsb/F/values=1024   27.0k ± 1%   27.6k ± 1%  +2.28%  (p=0.008 n=5+5)
ycsb/C/values=1024   1.08M ±12%   1.03M ±34%    ~     (p=0.310 n=5+5)
ycsb/B/values=1024    390k ± 8%    385k ± 9%    ~     (p=0.548 n=5+5)

name                old read     new read     delta
ycsb/E/values=1024    134G ± 5%    133G ± 5%    ~     (p=0.841 n=5+5)
ycsb/A/values=1024    513G ± 1%    511G ± 1%    ~     (p=0.222 n=5+5)
ycsb/D/values=1024    199G ±12%    196G ±14%    ~     (p=0.690 n=5+5)
ycsb/F/values=1024    508G ± 0%    505G ± 0%  -0.59%  (p=0.016 n=5+5)
ycsb/C/values=1024   63.8G ± 3%   64.1G ± 4%    ~     (p=1.000 n=5+5)
ycsb/B/values=1024    286G ± 6%    283G ± 7%    ~     (p=0.421 n=5+5)

name                old write    new write    delta
ycsb/E/values=1024    147G ± 5%    147G ± 6%    ~     (p=0.841 n=5+5)
ycsb/A/values=1024    606G ± 1%    605G ± 1%    ~     (p=0.548 n=5+5)
ycsb/D/values=1024    221G ±13%    217G ±14%    ~     (p=0.690 n=5+5)
ycsb/F/values=1024    612G ± 0%    612G ± 0%    ~     (p=0.095 n=5+5)
ycsb/C/values=1024   63.8G ± 3%   64.1G ± 4%    ~     (p=1.000 n=5+5)
ycsb/B/values=1024    326G ± 6%    323G ± 7%    ~     (p=0.421 n=5+5)

name                old r-amp    new r-amp    delta
ycsb/E/values=1024    8.79 ± 3%    8.87 ± 4%    ~     (p=0.452 n=5+5)
ycsb/A/values=1024    9.65 ± 6%    9.97 ±10%    ~     (p=0.690 n=5+5)
ycsb/D/values=1024    8.03 ± 4%    8.16 ± 4%    ~     (p=0.881 n=5+5)
ycsb/F/values=1024    0.00         0.00         ~     (all equal)
ycsb/C/values=1024    2.47 ±31%    2.46 ±47%    ~     (p=0.921 n=5+5)
ycsb/B/values=1024    6.28 ± 1%    6.33 ± 1%    ~     (p=0.095 n=5+5)

name                old w-amp    new w-amp    delta
ycsb/E/values=1024    21.7 ± 2%    21.8 ± 3%    ~     (p=0.548 n=5+5)
ycsb/A/values=1024    6.67 ± 1%    6.65 ± 1%    ~     (p=0.508 n=5+5)
ycsb/D/values=1024    20.2 ± 4%    20.4 ± 3%    ~     (p=1.000 n=5+5)
ycsb/F/values=1024    11.6 ± 1%    11.4 ± 1%  -2.34%  (p=0.016 n=5+5)
ycsb/C/values=1024    0.00         0.00         ~     (all equal)
ycsb/B/values=1024    8.58 ± 2%    8.60 ± 2%    ~     (p=0.897 n=5+5)
```

benchstat comparison for all ycsb benchmarks run for 30 minutes, on gce-n1-standard-16 machines configured with a large ssd.

```
name                old ops/sec  new ops/sec  delta
ycsb/E/values=1024   61.3k ± 3%   61.1k ± 2%     ~     (p=1.000 n=5+5)
ycsb/A/values=1024    189k ± 7%    188k ± 3%     ~     (p=0.690 n=5+5)
ycsb/D/values=1024    196k ± 5%    193k ± 4%     ~     (p=0.690 n=5+5)
ycsb/F/values=1024   73.7k ± 7%   66.3k ±11%  -10.00%  (p=0.032 n=5+5)
ycsb/C/values=1024    864k ±18%    916k ±20%     ~     (p=0.548 n=5+5)
ycsb/B/values=1024    353k ± 4%    344k ± 3%     ~     (p=0.222 n=5+5)

name                old read     new read     delta
ycsb/E/values=1024    110G ± 2%    112G ± 2%     ~     (p=0.222 n=5+5)
ycsb/A/values=1024   1.00T ± 7%   0.99T ± 3%     ~     (p=0.690 n=5+5)
ycsb/D/values=1024    311G ± 5%    308G ± 3%     ~     (p=0.841 n=5+5)
ycsb/F/values=1024   1.16T ±41%   0.86T ±28%     ~     (p=0.056 n=5+5)
ycsb/C/values=1024   50.3G ± 6%   51.9G ± 4%     ~     (p=0.310 n=5+5)
ycsb/B/values=1024    281G ± 4%    276G ± 2%     ~     (p=0.548 n=5+5)

name                old write    new write    delta
ycsb/E/values=1024    122G ± 2%    124G ± 3%     ~     (p=0.222 n=5+5)
ycsb/A/values=1024   1.19T ± 7%   1.18T ± 3%     ~     (p=0.690 n=5+5)
ycsb/D/values=1024    349G ± 5%    345G ± 3%     ~     (p=0.841 n=5+5)
ycsb/F/values=1024   1.45T ±34%   1.11T ±24%     ~     (p=0.056 n=5+5)
ycsb/C/values=1024   50.3G ± 6%   52.1G ± 6%     ~     (p=0.310 n=5+5)
ycsb/B/values=1024    318G ± 4%    312G ± 2%     ~     (p=0.548 n=5+5)

name                old r-amp    new r-amp    delta
ycsb/E/values=1024    7.78 ± 4%    8.06 ± 4%     ~     (p=0.151 n=5+5)
ycsb/A/values=1024    7.03 ± 1%    7.16 ± 2%     ~     (p=0.143 n=5+5)
ycsb/D/values=1024    6.91 ± 1%    7.01 ± 2%     ~     (p=0.143 n=5+5)
ycsb/F/values=1024    0.00         0.00          ~     (all equal)
ycsb/C/values=1024    2.79 ±21%    2.62 ±49%     ~     (p=1.000 n=5+5)
ycsb/B/values=1024    5.74 ± 1%    5.80 ± 1%     ~     (p=0.071 n=5+5)

name                old w-amp    new w-amp    delta
ycsb/E/values=1024    20.3 ± 2%    20.1 ±11%     ~     (p=1.000 n=5+5)
ycsb/A/values=1024    6.43 ± 0%    6.42 ± 0%     ~     (p=0.286 n=5+5)
ycsb/D/values=1024    18.2 ± 1%    18.1 ± 4%     ~     (p=0.690 n=5+5)
ycsb/F/values=1024    9.90 ±27%    8.48 ±14%     ~     (p=0.095 n=5+5)
ycsb/C/values=1024    0.00         0.00          ~     (all equal)
ycsb/B/values=1024    9.23 ± 1%    9.25 ± 1%     ~     (p=1.000 n=5+5)
```