@bananabrick 2022-01-27
Here's the changes from the microbenchmarks. Differences are trivial enough to not be a concern.

```
arjunnair@gceworker-arjunnair:~/go/src/github.com/cockroachdb/pebble$ benchstat t3 t6
name                                                          old time/op    new time/op    delta
Writer/block=4.0_K/filter=true/compression=NoCompression-24      136ms ± 2%     139ms ± 2%  +2.56%  (p=0.017 n=6+5)
Writer/block=4.0_K/filter=true/compression=Snappy-24             162ms ± 1%     165ms ± 1%  +1.91%  (p=0.004 n=6+5)
Writer/block=4.0_K/filter=true/compression=ZSTD-24               502ms ± 2%     509ms ± 1%    ~     (p=0.095 n=5+5)
Writer/block=4.0_K/filter=false/compression=NoCompression-24    85.9ms ± 1%    87.4ms ± 2%  +1.77%  (p=0.004 n=6+5)
Writer/block=4.0_K/filter=false/compression=Snappy-24            112ms ± 2%     114ms ± 2%    ~     (p=0.082 n=6+5)
Writer/block=4.0_K/filter=false/compression=ZSTD-24              452ms ± 6%     452ms ± 2%    ~     (p=0.429 n=6+5)
Writer/block=32_K/filter=true/compression=NoCompression-24       133ms ± 1%     134ms ± 0%    ~     (p=0.222 n=5+5)
Writer/block=32_K/filter=true/compression=Snappy-24              156ms ± 2%     154ms ± 1%    ~     (p=0.421 n=5+5)
Writer/block=32_K/filter=true/compression=ZSTD-24                239ms ± 0%     243ms ± 2%  +1.57%  (p=0.008 n=5+5)
Writer/block=32_K/filter=false/compression=NoCompression-24     81.7ms ± 1%    82.8ms ± 1%    ~     (p=0.082 n=6+5)
Writer/block=32_K/filter=false/compression=Snappy-24             103ms ± 1%     102ms ± 0%    ~     (p=0.247 n=6+5)
Writer/block=32_K/filter=false/compression=ZSTD-24               186ms ± 1%     186ms ± 0%    ~     (p=1.000 n=6+5)

name                                                          old speed      new speed      delta
Writer/block=4.0_K/filter=true/compression=NoCompression-24    291MB/s ± 2%   284MB/s ± 2%  -2.50%  (p=0.017 n=6+5)
Writer/block=4.0_K/filter=true/compression=Snappy-24          67.9MB/s ± 1%  66.7MB/s ± 1%  -1.87%  (p=0.004 n=6+5)
Writer/block=4.0_K/filter=true/compression=ZSTD-24            11.9MB/s ± 2%  11.8MB/s ± 1%    ~     (p=0.087 n=5+5)
Writer/block=4.0_K/filter=false/compression=NoCompression-24   446MB/s ± 1%   438MB/s ± 2%  -1.74%  (p=0.004 n=6+5)
Writer/block=4.0_K/filter=false/compression=Snappy-24         87.1MB/s ± 2%  86.0MB/s ± 2%    ~     (p=0.082 n=6+5)
Writer/block=4.0_K/filter=false/compression=ZSTD-24           10.5MB/s ± 6%  10.5MB/s ± 2%    ~     (p=0.429 n=6+5)
Writer/block=32_K/filter=true/compression=NoCompression-24     293MB/s ± 1%   292MB/s ± 0%    ~     (p=0.222 n=5+5)
Writer/block=32_K/filter=true/compression=Snappy-24           51.6MB/s ± 2%  52.1MB/s ± 1%    ~     (p=0.421 n=5+5)
Writer/block=32_K/filter=true/compression=ZSTD-24             14.3MB/s ± 0%  14.1MB/s ± 2%  -1.55%  (p=0.008 n=5+5)
Writer/block=32_K/filter=false/compression=NoCompression-24    463MB/s ± 1%   456MB/s ± 1%    ~     (p=0.082 n=6+5)
Writer/block=32_K/filter=false/compression=Snappy-24          65.9MB/s ± 1%  66.2MB/s ± 0%    ~     (p=0.247 n=6+5)
Writer/block=32_K/filter=false/compression=ZSTD-24            11.7MB/s ± 1%  11.7MB/s ± 0%    ~     (p=1.000 n=6+5)

name                                                          old alloc/op   new alloc/op   delta
Writer/block=4.0_K/filter=true/compression=NoCompression-24     25.1MB ± 0%    25.1MB ± 0%  +0.01%  (p=0.004 n=6+5)
Writer/block=4.0_K/filter=true/compression=Snappy-24            25.1MB ± 0%    25.1MB ± 0%  +0.02%  (p=0.004 n=6+5)
Writer/block=4.0_K/filter=true/compression=ZSTD-24              84.1MB ± 0%    84.1MB ± 0%  +0.02%  (p=0.008 n=5+5)
Writer/block=4.0_K/filter=false/compression=NoCompression-24     799kB ± 0%     800kB ± 0%  +0.10%  (p=0.004 n=6+5)
Writer/block=4.0_K/filter=false/compression=Snappy-24            829kB ± 0%     830kB ± 0%  +0.11%  (p=0.004 n=6+5)
Writer/block=4.0_K/filter=false/compression=ZSTD-24             59.8MB ± 0%    59.9MB ± 0%  +0.02%  (p=0.004 n=6+5)
Writer/block=32_K/filter=true/compression=NoCompression-24      25.2MB ± 0%    25.2MB ± 0%  +0.01%  (p=0.004 n=6+5)
Writer/block=32_K/filter=true/compression=Snappy-24             25.4MB ± 0%    25.5MB ± 0%  +0.01%  (p=0.004 n=6+5)
Writer/block=32_K/filter=true/compression=ZSTD-24               74.1MB ± 0%    74.1MB ± 0%  +0.01%  (p=0.008 n=5+5)
Writer/block=32_K/filter=false/compression=NoCompression-24      953kB ± 0%     954kB ± 0%  +0.07%  (p=0.004 n=6+5)
Writer/block=32_K/filter=false/compression=Snappy-24            1.16MB ± 0%    1.16MB ± 0%  +0.06%  (p=0.004 n=6+5)
Writer/block=32_K/filter=false/compression=ZSTD-24              49.8MB ± 0%    49.8MB ± 0%  +0.01%  (p=0.004 n=6+5)

name                                                          old allocs/op  new allocs/op  delta
Writer/block=4.0_K/filter=true/compression=NoCompression-24        146 ± 1%       154 ± 1%  +4.97%  (p=0.004 n=6+5)
Writer/block=4.0_K/filter=true/compression=Snappy-24               151 ± 0%       159 ± 2%  +5.30%  (p=0.008 n=5+5)
Writer/block=4.0_K/filter=true/compression=ZSTD-24               85.2k ± 0%     85.2k ± 0%  +0.03%  (p=0.008 n=5+5)
Writer/block=4.0_K/filter=false/compression=NoCompression-24       100 ± 0%       104 ± 0%  +4.00%  (p=0.008 n=5+5)
Writer/block=4.0_K/filter=false/compression=Snappy-24              106 ± 0%       110 ± 0%  +3.77%  (p=0.010 n=6+4)
Writer/block=4.0_K/filter=false/compression=ZSTD-24              85.2k ± 0%     85.2k ± 0%  +0.02%  (p=0.004 n=6+5)
Writer/block=32_K/filter=true/compression=NoCompression-24         140 ± 0%       148 ± 0%  +5.18%  (p=0.004 n=6+5)
Writer/block=32_K/filter=true/compression=Snappy-24                145 ± 0%       153 ± 0%  +5.00%  (p=0.004 n=6+5)
Writer/block=32_K/filter=true/compression=ZSTD-24                10.5k ± 0%     10.6k ± 0%  +0.15%  (p=0.004 n=6+5)
Writer/block=32_K/filter=false/compression=NoCompression-24       95.0 ± 0%      98.6 ± 1%  +3.79%  (p=0.004 n=6+5)
Writer/block=32_K/filter=false/compression=Snappy-24               100 ± 0%       103 ± 0%  +3.00%  (p=0.016 n=5+4)
Writer/block=32_K/filter=false/compression=ZSTD-24               10.5k ± 0%     10.5k ± 0%  +0.12%  (p=0.004 n=6+5)
```